+++ R/show_query.R
@@ -7,7 +7,8 @@
 show_query.tbl_lazy_json <- function(x, ...) {
   cli::cat_line(cli::style_bold("<Compute steps>"))
   purrr::walk(x$compute_steps, function(compute_step) {
-    cli::cat_line(glue::glue("[{compute_step$verb}], <{purrr::map(compute_step$params, ~ .)}>"))
+    params <- paste0(as.character(compute_step$params), collapse = ", ")
+    cli::cat_line(glue::glue("[{compute_step$verb}], <{params}>"))
   })
   invisible(x)
 }